refactor: refactor ui
