+++ ui/src/main/java/com/cairosquad/ui/auth/SignUpWebViewScreen.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.ui.login
+package com.cairosquad.ui.auth
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -6,8 +6,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.systemBarsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.R
-import androidx.compose.ui.res.stringResource
 import com.cairosquad.design_system.basic_component.AppBar
 import com.cairosquad.design_system.basic_component.WebView
 import com.cairosquad.ui.navigation.LocalNavController
