+++ ui/src/main/java/com/cairosquad/ui/navigation/AppNavigation.kt
@@ -21,9 +21,9 @@ import com.cairosquad.ui.details.TopCastScreen
 import com.cairosquad.ui.details.artist.ArtistScreen
 import com.cairosquad.ui.details.similar_movies.SimilarMoviesScreen
 import com.cairosquad.ui.details.similar_series.SimilarSeriesScreen
-import com.cairosquad.ui.login.ForgetPasswordWebViewScreen
-import com.cairosquad.ui.login.LoginScreen
-import com.cairosquad.ui.login.SignUpWebViewScreen
+import com.cairosquad.ui.auth.ForgetPasswordWebViewScreen
+import com.cairosquad.ui.auth.LoginScreen
+import com.cairosquad.ui.auth.SignUpWebViewScreen
 import com.cairosquad.ui.search.ForYouScreen
 import com.cairosquad.ui.splash.SplashScreen
